[ot] scripts/opentitan: verilate.py: split into front-end and modules
