+++ python/qemu/ot/verilator/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (c) 2025 Rivos, Inc.
+# SPDX-License-Identifier: Apache2
+
+"""OpenTitan Verilator tools
+
+   :author: Emmanuel Blot <eblot@rivosinc.com>
+"""
+
+DEFAULT_TIMEOUT = 10.0 * 60
+"""Default execution timeout."""
